Create README.md
The purpose of this project is to create an app that returns the air quality index in Polish cities based on real-time data from minitoring stations. The app is build with dash and classifies most recent available measurements into 6 categories according to the table below. There are 5 distinct pollutants taken into account when calculating AQI. Some monitoring stations have additional sensors measuring other pollutants. In these cases the app returns the information that the pollutant is not included in AQI index. The application uses 'monitoring_stations_PL.csv' dataset.